Reader: delete sites without search
