+++ config/horizon.json
@@ -77,7 +77,6 @@
 		"reader/following-manage-refresh": true,
 		"reader/related-posts": true,
 		"reader/search": true,
-		"reader/sites-in-search": true,
 		"resume-editing": true,
 		"republicize": true,
 		"rubberband-scroll-disable": false,
