+++ config/production.json
@@ -79,7 +79,6 @@
 		"reader/tags-with-elasticsearch": false,
 		"reader/related-posts": true,
 		"reader/search": true,
-		"reader/sites-in-search": true,
 		"resume-editing": true,
 		"republicize": true,
 		"rubberband-scroll-disable": false,
